HQ Function Organization
Made robot creation progs into functions that are made in the HQ
superclass for neatness and simplicity. The function is passed
a data of bit length 20, and a robot job of bit length 4.

Committed to allow for branching, to test A* method that might
or might not blow up the entire program.