Added pathfinding AI
Created an AI that finds a path to the destination using the
A* search algorithm. On caves it takes 50-60 rounds..
Big improvement from 200.
Also makes the HQ send out a message to calculate a path,
and makes the first medbay respond, "I got it!" and do it,
making the HQ shut up.