Added scaling to MapModule
This allows A* searches to be performed on a down-sampled
field, causing it to run faster but produce less accurate
results.

Also added an A* navigation routine to NavModule, which works
best when the module is given a series of closish waypoints
to follow.  It performs a limited A* search from waypoint to
waypoint.